@@ -13,18 +13,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "10.0.2",
-    "@angular/cdk": "^10.0.1",
-    "@angular/common": "10.0.2",
-    "@angular/compiler": "10.0.2",
-    "@angular/core": "10.0.2",
+    "@angular/animations": "10.1.1",
+    "@angular/cdk": "^10.2.1",
+    "@angular/common": "10.1.1",
+    "@angular/compiler": "10.1.1",
+    "@angular/core": "10.1.1",
     "@angular/flex-layout": "^10.0.0-beta.32",
-    "@angular/forms": "10.0.2",
-    "@angular/material": "^10.0.1",
-    "@angular/platform-browser": "10.0.2",
-    "@angular/platform-browser-dynamic": "10.0.2",
-    "@angular/router": "10.0.2",
-    "@angular/service-worker": "10.0.2",
+    "@angular/forms": "10.1.1",
+    "@angular/material": "^10.2.1",
+    "@angular/platform-browser": "10.1.1",
+    "@angular/platform-browser-dynamic": "10.1.1",
+    "@angular/router": "10.1.1",
+    "@angular/service-worker": "10.1.1",
     "@ngx-translate/core": "^13.0.0",
     "@ngx-translate/http-loader": "^6.0.0",
     "@stomp/ng2-stompjs": "^7.2.0",
@@ -43,10 +43,10 @@
     "zone.js": "~0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.1000.0",
-    "@angular/cli": "10.0.0",
-    "@angular/compiler-cli": "^10.0.2",
-    "@angular/language-service": "10.0.2",
+    "@angular-devkit/build-angular": "~0.1001.1",
+    "@angular/cli": "10.1.1",
+    "@angular/compiler-cli": "^10.1.1",
+    "@angular/language-service": "10.1.1",
     "@types/jasmine": "^3.5.11",
     "@types/jasminewd2": "~2.0.8",
     "@types/node": "^12.12.6",
