@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { ActivatedRouteSnapshot, Resolve, RouterStateSnapshot } from '@angular/router';
 import { Observable } from 'rxjs';
-import { map, flatMap } from 'rxjs/operators';
+import { map, mergeMap } from 'rxjs/operators';
 import { Comment } from '../models/comment';
 import { CommentService } from '../services/http/comment.service';
 import { RoomService } from '../services/http/room.service';
@@ -18,8 +18,8 @@ export class CommentResolver implements Resolve<Comment> {
 
   resolve(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<Comment> {
     return this.roomService.getRoomByShortId(route.params['shortId']).pipe(
-      flatMap(room => {
-        return this.commentService.getComment(route.params['commentId'], room.id)
+      mergeMap(room => {
+        return this.commentService.getComment(route.params['commentId'], room.id);
       })
     );
   }
