@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { ArsComponent } from './ars.component';
 
@@ -6,7 +6,7 @@ describe('ArsComponent', () => {
   let component: ArsComponent;
   let fixture: ComponentFixture<ArsComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ ArsComponent ]
     })
