@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import {  ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { ArsSliderCombComponent } from './ars-slider-comb.component';
 
@@ -6,7 +6,7 @@ describe('ArsSliderCombComponent', () => {
   let component: ArsSliderCombComponent;
   let fixture: ComponentFixture<ArsSliderCombComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ ArsSliderCombComponent ]
     })
