@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { RespComponent } from './resp.component';
 
@@ -6,7 +6,7 @@ describe('RespComponent', () => {
   let component: RespComponent;
   let fixture: ComponentFixture<RespComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ RespComponent ]
     })
