@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { FillComponent } from './fill.component';
 
@@ -6,7 +6,7 @@ describe('FillComponent', () => {
   let component: FillComponent;
   let fixture: ComponentFixture<FillComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ FillComponent ]
     })
