@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { FrameTestComponent } from './frame-test.component';
 
@@ -6,7 +6,7 @@ describe('FrameTestComponent', () => {
   let component: FrameTestComponent;
   let fixture: ComponentFixture<FrameTestComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ FrameTestComponent ]
     })
