@@ -1,4 +1,4 @@
-/* import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+/* import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { ContentCreationPageComponent } from './content-creation-page.component';
 
@@ -6,7 +6,7 @@ describe('ContentCreationPageComponent', () => {
   let component: ContentCreationPageComponent;
   let fixture: ComponentFixture<ContentCreationPageComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ ContentCreationPageComponent ]
     })
