@@ -1,5 +1,5 @@
 /*
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { ContentSlideCreationComponent } from './content-slide-creation.component';
 
@@ -7,7 +7,7 @@ describe('ContentSlideCreationComponent', () => {
   let component: ContentSlideCreationComponent;
   let fixture: ComponentFixture<ContentSlideCreationComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ ContentSlideCreationComponent ]
     })
