@@ -1,4 +1,4 @@
-/* import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+/* import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { ContentListComponent } from './content-list.component';
 
@@ -6,7 +6,7 @@ describe('ContentListComponent', () => {
   let component: ContentListComponent;
   let fixture: ComponentFixture<ContentListComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ ContentListComponent ]
     })
