@@ -1,4 +1,4 @@
-/* import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+/* import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { RoomCreatorPageComponent } from './room-creator-page.component';
 
@@ -6,7 +6,7 @@ describe('RoomCreatorPageComponent', () => {
   let component: RoomCreatorPageComponent;
   let fixture: ComponentFixture<RoomCreatorPageComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ RoomCreatorPageComponent ]
     })
