@@ -1,4 +1,4 @@
-/*import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+/*import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { CommentSettingsComponent } from './comment-settings.component';
 
@@ -6,7 +6,7 @@ describe('CommentSettingsComponent', () => {
   let component: CommentSettingsComponent;
   let fixture: ComponentFixture<CommentSettingsComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ CommentSettingsComponent ]
     })
