@@ -1,5 +1,5 @@
 /**
- import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+ import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
  import { CookiesComponent } from './cookies.component';
 
@@ -7,7 +7,7 @@
   let component: CookiesComponent;
   let fixture: ComponentFixture<CookiesComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ CookiesComponent ]
     })
