@@ -1,5 +1,5 @@
 /**
- import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+ import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
  import { OverlayComponent } from './overlay.component';
 
@@ -7,7 +7,7 @@
   let component: OverlayComponent;
   let fixture: ComponentFixture<OverlayComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ OverlayComponent ]
     })
