@@ -1,4 +1,4 @@
-/* import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+/* import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { ContentTextParticipantComponent } from './content-text-participant.component';
 
@@ -6,7 +6,7 @@ describe('ContentTextParticipantComponent', () => {
   let component: ContentTextParticipantComponent;
   let fixture: ComponentFixture<ContentTextParticipantComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ ContentTextParticipantComponent ]
     })
