@@ -1,4 +1,4 @@
-/* import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+/* import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { ParticipantContentCarouselPageComponent } from './participant-content-carousel-page.component';
 
@@ -6,7 +6,7 @@ describe('CreatorContentCarouselPageComponent', () => {
   let component: ParticipantContentCarouselPageComponent;
   let fixture: ComponentFixture<ParticipantContentCarouselPageComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ ParticipantContentCarouselPageComponent ]
     })
