@@ -1,4 +1,4 @@
-/* import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+/* import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { RoomParticipantPageComponent } from './room-participant-page.component';
 
@@ -6,7 +6,7 @@ describe('RoomParticipantPageComponent', () => {
   let component: RoomParticipantPageComponent;
   let fixture: ComponentFixture<RoomParticipantPageComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ RoomParticipantPageComponent ]
     })
