@@ -1,4 +1,4 @@
-/** import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+/** import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
  import { CommentExportComponent } from './comment-export.component';
 
@@ -6,7 +6,7 @@
   let component: CommentExportComponent;
   let fixture: ComponentFixture<CommentExportComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ CommentExportComponent ]
     })
