@@ -1,12 +1,12 @@
 /*
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 import { PresentCommentComponent } from './present-comment.component';
 
 describe('PresentCommentComponent', () => {
   let component: PresentCommentComponent;
   let fixture: ComponentFixture<PresentCommentComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ PresentCommentComponent ]
     })
