@@ -1,5 +1,5 @@
 /*
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { UpdateInfoComponent } from './update-info.component';
 
@@ -7,7 +7,7 @@ describe('UpdateInfoComponent', () => {
   let component: UpdateInfoComponent;
   let fixture: ComponentFixture<UpdateInfoComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ UpdateInfoComponent ]
     })
