@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { CommentPageComponent } from './comment-page.component';
 import { Injectable, Renderer2, Component, Input } from '@angular/core';
@@ -87,7 +87,7 @@ describe('CommentPageComponent', () => {
   let component: CommentPageComponent;
   let fixture: ComponentFixture<CommentPageComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [
         CommentPageComponent,
