@@ -1,4 +1,4 @@
-/* import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+/* import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { RenderedTextComponent } from './rendered-text.component';
 
@@ -6,7 +6,7 @@ describe('RenderedTextComponent', () => {
   let component: RenderedTextComponent;
   let fixture: ComponentFixture<RenderedTextComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ RenderedTextComponent ]
     })
