@@ -1,5 +1,5 @@
 /*
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { StatisticChoiceComponent } from './statistic-choice.component';
 
@@ -7,7 +7,7 @@ describe('StatisticChoiceComponent', () => {
   let component: StatisticChoiceComponent;
   let fixture: ComponentFixture<StatisticChoiceComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ StatisticChoiceComponent ]
     })
