@@ -1,4 +1,4 @@
-/* import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+/* import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { StatisticsPageComponent } from './statistics-page.component';
 
@@ -6,7 +6,7 @@ describe('StatisticsPageComponent', () => {
   let component: StatisticsPageComponent;
   let fixture: ComponentFixture<StatisticsPageComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ StatisticsPageComponent ]
     })
