@@ -1,4 +1,4 @@
-import { UserRole } from './user-roles.enum'
+import { UserRole } from './user-roles.enum';
 
 export class RoomAccess {
   roomId: string;
